+++ framework/react/test-utils/CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ### Patch Changes
 
+- Updated dependencies
+  - @starbeam/core@0.5.4
+  - @starbeam/debug@0.5.4
+  - @starbeam/timeline@0.5.4
+  - @starbeam/core-utils@0.5.4
+  - @starbeam/peer@1.0.4
+  - @starbeam-workspace/test-utils@0.0.3
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [194e461]
 - Updated dependencies [d21391d]
   - @starbeam/core@0.5.3