+++ framework/react/use-strict-lifecycle/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @starbeam/use-strict-lifecycle
 
+## 0.5.4
+
+### Patch Changes
+
+- The 0.5.3 release was missing dists
+
 ## 0.5.3
 
 ### Patch Changes