+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @starbeam/core
 
+## 0.5.4
+
+### Patch Changes
+
+- The 0.5.3 release was missing dists
+- Updated dependencies
+  - @starbeam/debug@0.5.4
+  - @starbeam/timeline@0.5.4
+  - @starbeam/verify@0.5.4
+  - @starbeam/core-utils@0.5.4
+  - @starbeam/peer@1.0.4
+
 ## 0.5.3
 
 ### Patch Changes