+++ packages/debug/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @starbeam/debug
 
+## 0.5.4
+
+### Patch Changes
+
+- The 0.5.3 release was missing dists
+- Updated dependencies
+  - @starbeam/verify@0.5.4
+  - @starbeam/peer@1.0.4
+
 ## 0.5.3
 
 ### Patch Changes