+++ packages/peer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @starbeam/peer
 
+## 1.0.4
+
+### Patch Changes
+
+- The 0.5.3 release was missing dists
+
 ## 1.0.3
 
 ### Patch Changes